+++ ng2-smart-table.ts
@@ -1,7 +1,0 @@
-export * from './src/ng2-smart-table.directives';
-export * from './src/ng2-smart-table/ng2-smart-table.component';
-export * from './src/ng2-smart-table/lib';
-export * from './src/ng2-smart-table.module';
-export { Cell } from './src/ng2-smart-table/lib/data-set/cell';
-export { DefaultEditor, Editor } from './src/ng2-smart-table/components/cell/cell-editors/default-editor';
-export { ViewCell } from './src/ng2-smart-table/components/cell/cell-view-mode';