+++ src/app/pages/examples/custom-edit-view/basic-example-button-view.component.ts
@@ -1,0 +1,100 @@
+import { Component, OnInit, Input } from '@angular/core';
+
+import { ViewCell } from '../../../../ng2-smart-table';
+
+@Component({
+  selector: 'button-view',
+  template: `
+    <button (click)="showAlert()">{{ renderValue }}</button>
+  `,
+})
+export class ButtonViewComponent implements ViewCell, OnInit {
+  renderValue: string;
+
+  @Input() value: string | number;
+
+  constructor() { }
+
+  ngOnInit() {
+    this.renderValue = this.value.toString().toUpperCase();
+  }
+
+  showAlert() {
+    alert(this.renderValue);
+  }
+}
+
+@Component({
+  selector: 'basic-example-button-view',
+  template: `
+    <ng2-smart-table [settings]="settings" [source]="data"></ng2-smart-table>
+  `,
+})
+export class BasicExampleButtonViewComponent implements OnInit {
+
+  settings = {
+    columns: {
+      id: {
+        title: 'ID',
+      },
+      name: {
+        title: 'Full Name',
+      },
+      username: {
+        title: 'User Name',
+      },
+      email: {
+        title: 'Email',
+      },
+      button: {
+        title: 'Button',
+        type: 'custom',
+        renderComponent: ButtonViewComponent,
+      },
+    },
+  };
+
+  data = [
+    {
+      id: 1,
+      name: 'Leanne Graham',
+      username: 'Bret',
+      email: 'Sincere@april.biz',
+      button: 'Button #1',
+    },
+    {
+      id: 2,
+      name: 'Ervin Howell',
+      username: 'Antonette',
+      email: 'Shanna@melissa.tv',
+      button: 'Button #2',
+    },
+    {
+      id: 3,
+      name: 'Clementine Bauch',
+      username: 'Samantha',
+      email: 'Nathan@yesenia.net',
+      button: 'Button #3',
+    },
+    {
+      id: 4,
+      name: 'Patricia Lebsack',
+      username: 'Karianne',
+      email: 'Julianne.OConner@kory.org',
+      button: 'Button #4',
+    },
+    {
+      id: 5,
+      name: 'Chelsey Dietrich',
+      username: 'Kamren',
+      email: 'Lucio_Hettinger@annie.ca',
+      button: 'Button #5',
+    },
+  ];
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}