+++ src/ng2-smart-table/components/filter/filter-types/checkbox-filter.component.ts
@@ -3,13 +3,15 @@ import { FormControl } from '@angular/forms';
 
 import { DefaultFilter } from './default-filter';
 import { Column } from '../../../lib/data-set/column';
+import 'rxjs/add/operator/debounceTime';
 
 @Component({
   selector: 'checkbox-filter',
   template: `
     <input type="checkbox" [formControl]="inputControl" [ngClass]="inputClass" class="form-control">
-    <a href="#" *ngIf="filterActive" (click)="resetFilter($event)">{{column.getFilterConfig()?.resetText || 'reset'}}</a>
-  `
+    <a href="#" *ngIf="filterActive"
+                (click)="resetFilter($event)">{{column.getFilterConfig()?.resetText || 'reset'}}</a>
+  `,
 })
 export class CheckboxFilterComponent extends DefaultFilter implements OnInit {
 
@@ -32,7 +34,7 @@ export class CheckboxFilterComponent extends DefaultFilter implements OnInit {
       });
   }
 
-  resetFilter(event) {
+  resetFilter(event: any) {
     event.preventDefault();
     this.query = '';
     this.inputControl.setValue(false, { emitEvent: false });