+++ src/ng2-smart-table/components/filter/filter-types/select-filter.component.ts
@@ -3,6 +3,8 @@ import { FormControl } from '@angular/forms';
 
 import { DefaultFilter } from './default-filter';
 import { Column } from '../../../lib/data-set/column';
+import 'rxjs/add/operator/distinctUntilChanged';
+import 'rxjs/add/operator/debounceTime';
 
 @Component({
   selector: 'select-filter',
@@ -17,7 +19,7 @@ import { Column } from '../../../lib/data-set/column';
           {{ option.title }}
         </option>
     </select>
-  `
+  `,
 })
 export class SelectFilterComponent extends DefaultFilter implements OnInit {