+++ src/ng2-smart-table/components/tbody/cells/index.ts
@@ -1,2 +1,0 @@
-export * from './edit-delete.component';
-export * from './create-cancel.component';