+++ src/ng2-smart-table/components/tbody/tbody.module.ts
@@ -1,0 +1,30 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { FormsModule } from '@angular/forms';
+
+import { CellModule } from '../cell/cell.module';
+
+import { Ng2SmartTableTbodyComponent } from './tbody.component';
+import { TbodyCreateCancelComponent } from './cells/create-cancel.component';
+import { TbodyEditDeleteComponent } from './cells/edit-delete.component';
+
+const TBODY_COMPONENTS = [
+  TbodyCreateCancelComponent,
+  TbodyEditDeleteComponent,
+  Ng2SmartTableTbodyComponent,
+];
+
+@NgModule({
+  imports: [
+    CommonModule,
+    FormsModule,
+    CellModule,
+  ],
+  declarations: [
+    ...TBODY_COMPONENTS,
+  ],
+  exports: [
+    ...TBODY_COMPONENTS,
+  ],
+})
+export class TBodyModule { }