+++ CHANGELOG.md
@@ -1,11 +1,47 @@
+<a name="1.0.2"></a>
+# 1.0.2 (2017-04-04)
+
+### Bug fixes
+
+* **build:** move ng2-completer and lodash from peerDependencies to dependencies
+
+<a name="1.0.1"></a>
+# 1.0.1 (2017-03-29)
+
+### Bug fixes
+
+* **build:** fix package.json
+
+<a name="1.0.0"></a>
+# 1.0.0 (2017-03-29)
+
+### Features
+
+* **table:** update to angular 4.0.0
+* **demo:** @angular/cli integration
+* **table:** add a row hover event ([68b73ee](https://github.com/akveo/ng2-smart-table/commit/68b73eef0b4607cef522f201c3e67a230a5ee03e)), ([1ceca0d](https://github.com/akveo/ng2-smart-table/commit/1ceca0d6afa3be3ba5f2ef53447d9a75f360150a)), ([#253](https://github.com/akveo/ng2-smart-table/pull/253)), ([#262](https://github.com/akveo/ng2-smart-table/pull/262))
+* **table:** add a deselect row ([2ab02b4](https://github.com/akveo/ng2-smart-table/commit/2ab02b4949c577da34153f7ad1870e7acaa399e6)), ([#94](https://github.com/akveo/ng2-smart-table/pull/94))
+* **table:** add a getFilteredAndSorted function ([df3f343](https://github.com/akveo/ng2-smart-table/commit/df3f343caaff9aff1ecc6d8b94767e3466524068)), ([#249](https://github.com/akveo/ng2-smart-table/pull/249))
+
+### Bug fixes
+
+* **table:** ng2-completer failed with empty value ([b0e1797](https://github.com/akveo/ng2-smart-table/commit/b0e1797e9d22ca4a076e01120525f16628e3e6a6)), closes ([#164](https://github.com/akveo/ng2-smart-table/issues/164))
+* **table:** add a check for existing filters ([8daa70f](https://github.com/akveo/ng2-smart-table/commit/8daa70f8434cfb4036cec07dc68208a3c3c51d00)), ([#40](https://github.com/akveo/ng2-smart-table/pull/40))
+* **table:** fix the actions column positioning ([ebfbd54](https://github.com/akveo/ng2-smart-table/commit/ebfbd5424a227217484494fc48a5f2fb854a470d)), ([#245](https://github.com/akveo/ng2-smart-table/pull/245))
+
+### Code Refactoring
+
+* **table:** move the components to the modules
+* **demo:** move some examples to new /examples route ([#213](https://github.com/akveo/ng2-smart-table/pull/213))
+
 <a name="0.6.0-0"></a>
 # 0.6.0-0 (2017-03-02)
 
 ### Features
 
-* **core:** configuring column filters ([#185](https://github.com/akveo/ng2-smart-table/pull/185))
-* **core:** custom cell view component ([#193](https://github.com/akveo/ng2-smart-table/pull/193))
-* **core:** update to angular 2.4.9
+* **table:** configuring column filters ([#185](https://github.com/akveo/ng2-smart-table/pull/185))
+* **table:** custom cell view component ([#193](https://github.com/akveo/ng2-smart-table/pull/193))
+* **table:** update to angular 2.4.9
 
 ### Bug Fixes
 
@@ -13,14 +49,14 @@
 
 ### Code Refactoring
 
-* **core:** grouping components by section ([#172](https://github.com/akveo/ng2-smart-table/pull/172))
+* **table:** grouping components by section ([#172](https://github.com/akveo/ng2-smart-table/pull/172))
 
 <a name="0.3.1"></a>
 # 0.3.1 (2016-09-19)
 
 ### Features
 
-* **core:** update to angular 2.0
+* **table:** update to angular 2.0
 
 ### How to update
 
@@ -32,7 +68,7 @@
 
 ### Features
 
-* **core:** update to angular rc.5
+* **table:** update to angular rc.5
 
 ### How to update